// src/app/shared/models/User.ts
// DTO User
export class User {
    id: string;
    firstName: string;
    lastName: string;
    email: string;
    organization: string;
    phone: string;
    password: string;
    enabled: string;
    apiKey: string;
    organisationId: string;
    organisationName: string;
    isAdmin: boolean;
}